pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "pydicom>=3.0.1",
    "scikit-image>=0.25.2",
    "simpleitk>=2.5.0",
    "streamlit>=1.44.1",
]
